// libs/features/contributor/src/lib/shared/enums/engagement-sub-type.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum EngagementSubType {
  BACKDATED = 'ENGAGEMENT_WITH_BACKDATED',
  PENDING = 'ENGAGEMENT_WITH_PENDING',
  PENDING_TEMRINATION = 'ENGAGEMENT_WITH_PENDING_TERMINATION'
}