// libs/features/contributor/src/lib/shared/enums/nationalities.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

/* Nationalities enum */
export enum Nationalities {
  UAE = 'United Arab Emirates',
  BAHRAIN = 'Bahrain',
  QATAR = 'Qatar',
  OMAN = 'Oman',
  KUWAIT = 'Kuwait',
  SAUDI_ARABIA = 'Saudi Arabia',
  MIXED = 'Mixed',
  IMMIGRATED_TRIBE = 'Immigrated Tribes',
  NON_SAUDI = 'Non Saudi'
}