// libs/features/contributor/src/lib/shared/models/submit-engagement-response.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

import { BilingualText } from '@gosi-ui/core';

export class SubmitEngagementResponse {
  referenceNo: [] = [];
  message: BilingualText = new BilingualText();
}